Move the timeline status bar slider to the jupyter-collaboration-ui package
