+++ packages/collaboration-extension/package.json
@@ -61,8 +61,10 @@
     "@jupyterlab/apputils": "^4.7.0-beta.1",
     "@jupyterlab/codemirror": "^4.6.0-beta.1",
     "@jupyterlab/coreutils": "^6.6.0-beta.1",
+    "@jupyterlab/docregistry": "^4.6.0-beta.1",
     "@jupyterlab/services": "^7.6.0-beta.1",
     "@jupyterlab/statedb": "^4.6.0-beta.1",
+    "@jupyterlab/statusbar": "^4.6.0-beta.1",
     "@jupyterlab/translation": "^4.6.0-beta.1",
     "@jupyterlab/ui-components": "^4.6.0-beta.1",
     "@lumino/widgets": "^2.8.0",
