+++ packages/collaboration/package.json
@@ -41,6 +41,7 @@
   "dependencies": {
     "@codemirror/state": "^6.2.0",
     "@codemirror/view": "^6.7.0",
+    "@jupyter/docprovider": "^5.0.0-alpha.0",
     "@jupyterlab/apputils": "^4.7.0-beta.1",
     "@jupyterlab/coreutils": "^6.6.0-beta.1",
     "@jupyterlab/docregistry": "^4.6.0-beta.1",
