+++ packages/collaboration/src/index.ts
@@ -10,5 +10,6 @@ export * from './collaboratorspanel';
 export * from './cursors';
 export * from './menu';
 export * from './sharedlink';
+export * from './TimelineSlider';
 export * from './userinfopanel';
 export * from './users-item';
