+++ packages/collaboration/style/base.css
@@ -7,6 +7,7 @@
 @import url('./sidepanel.css');
 @import url('./users-item.css');
 @import url('./sharedlink.css');
+@import url('./slider.css');
 
 .jp-shared-link-body {
     user-select: none;
