+++ packages/collaboration/style/slider.css
