+++ packages/docprovider/src/index.ts
@@ -12,6 +12,5 @@ export * from './notebookCellExecutor';
 export * from './requests';
 export * from './ydrive';
 export * from './yprovider';
-export * from './TimelineSlider';
 export * from './tokens';
 export * from './forkManager';
