+++ packages/docprovider/style/base.css
@@ -2,5 +2,3 @@
 | Copyright (c) Jupyter Development Team.
 | Distributed under the terms of the Modified BSD License.
 |---------------------------------------------------------------------------- */
-
-@import url('./slider.css');
